HPCC-25736 Refactor React Fields, Groups and Forms
